attack/patch_attack/SemSegAdvPatch/requirements_gpu1.txt
backcall==0.2.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
imageio==2.9.0
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1

matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.1.48
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.18
protobuf==3.15.6
ptyprocess==0.7.0
pydensecrf==1.0rc3
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
scipy==0.19.0
six==1.15.0
tensorboardX==2.1

-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111

tornado==6.1
tqdm==4.11.2
traitlets==4.3.3
typing-extensions==3.7.4.3
wcwidth==0.2.5

kornia==0.5.1
